fix betty style, en todas las funciones, se creo una nueva funcion para liberar la memoria cuando se ejecute el comando exit, y se agrego un nuevo file para almacenar una funcion str, Pendiente eliminar 8 lineas del prompt
