+++ holberton.h
@@ -42,4 +42,5 @@ void _fork_fail(void);
 void free_list(list_t *head);
 char *_strdup(char *str);
 char *_strchr(char *s, char c);
+void free_exit(char *_path, char *data, list_t *head, char **token);
 #endif
